requirements.txt
evostra==1.0.1
gym==0.10.4
keras==2.1.5
numpy==1.14.2
Pillow==5.0.0
scikit-image==0.13.1
tensorflow==1.6.0
http://download.pytorch.org/whl/torch-0.3.1-cp36-cp36m-macosx_10_7_x86_64.whl ; sys_platform == 'darwin'
http://download.pytorch.org/whl/cpu/torch-0.3.1-cp36-cp36m-linux_x86_64.whl ; sys_platform == 'linux'